Reduce application structure approaches to only ReaderT and Run
The issue here is understandability. I can either re-implement each 
approach so that it is self-contained (all files are in the same folder) 
at the cost of having duplicate code that makes the ToC file much 
longer, or I can reduce duplicate code at the cost of making code a bit 
harder to understand for new learners.
By reducing the approaches to only the two approaches typically used 
(ReaderT design pattern and Run), duplicate code isn't as big of an 
issue and the ToC is much shorter than otherwise.