Reduce duplicate code; improve Free's API file
Free's API file didn't really show how it was connecting the pure domain logic with the impure effects that make it useful. This has now been corrected.

Moreover, the structure of Free's language wasn't efficient. One had to 'translate' the Free API language into the Terminal Query language rather than reuse the Free language as the Terminal language. This was also fixed in this update.