Update 12-Useful-Monads.md
I have other things to suggest, but I think I will open first an issue to discuss those.
This PR is literally fixing a bug on the JS code. The code you had previously outputs `["000","001","002","0010"...]` which is not the intention.
That's because for in is an iterator of the object properties, not array contents.
Also, there is no method called add on javascript, there is push which returns the index of the inserted value not a new list.