Further refine ToC generation tool
Main Changes:
- This approach will read the content of `Summary-header.md`, append the 
generated ToC contents to it, and write the result into `./SUMMARY.md`.
- This commit also ensures a folder appears in the resulting ToC. The 
link is grayed out as `mdbook` assumes that the file is a WIP. I'm 
exploiting it to force the ToC structure to mirror the repo's structure.
- The numeric prefix to a file or folder (e.g. `00 ReadMe.md`) no longer 
has that prefix. This was done via `cleanString`.

Future work:
- a folder that has a `ReadMe.md` file should use that file as its link.
- a `.purs` or `.js` file should have a file created for it in the 
`mdbook/files` folder. The content of the file should be:
---------
# File-Name

```purescript
{{#include file-name.purs}}
```
---------